email-assistant/pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "AI-powered email assistant that categorizes inbox and writes email drafts using agents"
authors = [
    {name = "AI Builders Bootcamp", email = "example@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.21.0",
    "python-dotenv>=1.0.0",
    "pandas>=1.5.0",
    "email-validator>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
